fixed bug in stack.c
I took a deeper look in the getop() function and compared it to the original one in K&R. An alternative implementation to keep the structure well is to ungetch(c) after assigning getch() to c in the else block of "if (c == '-')". I'm not sure exactly why that is, but I assume it's because after the first c input is assigned, the second one is missed out. E.g if you type '1200' in the program, the 2 will be skipped. This bug specifically targeted the second read digit.